Build and test x3270 without DBCS. Build and test x3270 with iconv. Allow build with -Wsigned-compare.
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@138 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/ansi.c
@@ -1792,7 +1792,7 @@ ansi_send_pf(int nn)
 		35, 36, 37, 38
 	};
 
-	if (nn < 1 || nn > sizeof(code)/sizeof(code[0]))
+	if (nn < 1 || (unsigned)nn > sizeof(code)/sizeof(code[0]))
 		return;
 	(void) sprintf(fn_buf, "\033[%d~", code[nn-1]);
 	net_sends(fn_buf);
